Shortcuts are stored in a central table, not in the widgets. This provides the following benefits:
  1. More than one shortcut per widget
  2. The same "best" widget is always found, rather than the
     first possible match, when searching.
  3. Programs can read the table and present a correct and readable list
     of the shortcut assignments, and can remove shortcuts from conflicting
     widgets.
  4. 4 bytes shaved off size of widget

Current implementation is simple, it should be improved to be a hash table
based on widget address. It should also be changed to store arbitrary
sparse data, such as needed by signal/slot schemes.

Other changes:

Group::focus() renamed to Group::focus_index() to avoid hiding fltk::focus().

Several indexes changed from int to unsigned.

Win32 "defers" window mapping/unmapping/destruction until it is ready to
call GetMessage. This should fix X programs that don't expect draw() to be
called by the constructors and destructors for windows.

Win32 paste would delete the text buffer pointed to by event_text().


git-svn-id: file:///fltk/svn/trunk@3887 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_get_key_mac.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_get_key_mac.cxx,v 1.7 2003/09/06 22:37:36 spitzak Exp $"
+// "$Id: Fl_get_key_mac.cxx,v 1.8 2004/11/12 06:50:17 spitzak Exp $"
 //
 // MacOS keyboard state routines for the Fast Light Tool Kit (FLTK).
 //
@@ -69,7 +69,7 @@ static const struct {unsigned short vk, fltk;} vktab[] = {
   { 58, LeftAltKey }, { 58, RightAltKey },
 };
 
-static int fltk2mac(int fltk) {
+static unsigned short fltk2mac(unsigned fltk) {
   if (fltk >= 'a' && fltk <= 'z') fltk-=('a'-'A');
   int a = 0;
   int b = sizeof(vktab)/sizeof(*vktab);
@@ -82,14 +82,14 @@ static int fltk2mac(int fltk) {
 }
 
 //: returns true, if that key was pressed during the last event
-bool fltk::event_key_state(int k) {
+bool fltk::event_key_state(unsigned k) {
   return get_key_state(k);
 }
 
 //#include <stdio.h>
 
 //: returns true, if that key is pressed right now
-bool fltk::get_key_state(int k) {
+bool fltk::get_key_state(unsigned k) {
   KeyMap foo;
   GetKeys(foo);
 #ifdef MAC_TEST_FOR_KEYCODES
@@ -105,5 +105,5 @@ bool fltk::get_key_state(int k) {
 }
 
 //
-// End of "$Id: Fl_get_key_mac.cxx,v 1.7 2003/09/06 22:37:36 spitzak Exp $".
+// End of "$Id: Fl_get_key_mac.cxx,v 1.8 2004/11/12 06:50:17 spitzak Exp $".
 //
